+++ Sources/Toml/Helpers.swift
@@ -90,7 +90,7 @@ func trimStringIdentifier(_ input: String, _ quote: String = "\"") -> String {
     let matches = regex.matches(in: input, options: [],
         range: NSMakeRange(0, input.utf16.count))
     let nss = (input as NSString)
-    return nss.substring(with: matches[0].rangeAt(1))
+    return nss.substring(with: matches[0].range(at: 1))
 }
 
 func getKeyPathFromTable(tokens: [Token]) -> [String] {