+++ Sources/Toml/Regex.swift
@@ -30,7 +30,7 @@ extension String {
         }
         
         let range = expression.rangeOfFirstMatch(in: self, options: [],
-            range: NSMakeRange(0, self.characters.count))
+            range: NSMakeRange(0, self.count))
         if range.location != NSNotFound {
             return (self as NSString).substring(with: range)
         }