+++ CHANGELOG.md
@@ -10,9 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - SetValueProtocol method renamed from func setValue(key: , value: ) to 
   func set(value:, for:)
 - Add support for running tests from XCode
-- Updated to Swift 3.0 GM
+- Updated to Swift 4.0.3
 - Updated examples in README.md to reflect current API
 - Significant performance improvements
+- Update String handling to not use .characters view per Swift4 conventions
+- Use range(at:) instead of rangeAt(_:)
 
 ## [0.4.0] - 2016-08-24
 ### Added